Fix CI: Ray.Di 2.21 rename and Compiler complexity
Ray.Di 2.21 defers rename() until after configure()+merge, so
CliModule's rename-then-bind moved CliRouter off RouterInterface.
Move the parent binding on lastModule first, then bind CliRouter.

Extract PreloadClassFilter to satisfy PHPMD, and construct it before
unregistering Composer autoload so the filter class can load.
