+++ src/Compiler/PreloadClassFilter.php
@@ -1,0 +1,134 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BEAR\Package\Compiler;
+
+use BEAR\Package\Compiler;
+use BEAR\Package\Provide\Error\NullPage;
+use Composer\Autoload\ClassLoader;
+use ReflectionClass;
+
+use function assert;
+use function class_exists;
+use function dirname;
+use function file_exists;
+use function interface_exists;
+use function is_int;
+use function is_string;
+use function realpath;
+use function str_starts_with;
+use function strpos;
+use function trait_exists;
+
+use const DIRECTORY_SEPARATOR;
+
+/**
+ * Filters declared symbols for preload.php generation.
+ */
+final class PreloadClassFilter
+{
+    private string $composerDir = '';
+
+    /** @var array<string, true> */
+    private array $composerLoadedFiles = [];
+
+    public function __construct(
+        private ClassLoader $loader,
+    ) {
+        $this->prepareComposerFiles();
+    }
+
+    public function __invoke(string $class): bool
+    {
+        if ($this->isExcludedClass($class) || ! $this->isDeclared($class)) {
+            return false;
+        }
+
+        /** @var class-string $class */
+        $reflection = new ReflectionClass($class);
+        if ($reflection->isAnonymous()) {
+            return false;
+        }
+
+        $filePath = $this->resolvedAutoloadPath($class, $reflection);
+        if ($filePath === null) {
+            return false;
+        }
+
+        return ! $this->isComposerLoadedFile($filePath);
+    }
+
+    private function isDeclared(string $class): bool
+    {
+        return class_exists($class, false) || interface_exists($class, false) || trait_exists($class, false);
+    }
+
+    /**
+     * @param class-string            $class
+     * @param ReflectionClass<object> $reflection
+     */
+    private function resolvedAutoloadPath(string $class, ReflectionClass $reflection): string|null
+    {
+        $fileName = $reflection->getFileName();
+        $loaderFile = $this->loader->findFile($class);
+        if (! is_string($fileName) || ! is_string($loaderFile)) {
+            return null;
+        }
+
+        $filePath = $this->normalizePath($fileName);
+        $loaderPath = $this->normalizePath($loaderFile);
+        if (! is_string($filePath) || $filePath !== $loaderPath) {
+            return null;
+        }
+
+        return $filePath;
+    }
+
+    public function isExcludedClass(string $class): bool
+    {
+        return $class === NullPage::class
+            || is_int(strpos($class, Compiler::class))
+            || is_int(strpos($class, NullPage::class));
+    }
+
+    private function prepareComposerFiles(): void
+    {
+        $classLoaderFile = (new ReflectionClass(ClassLoader::class))->getFileName();
+        assert(is_string($classLoaderFile));
+        $composerDir = $this->normalizePath(dirname($classLoaderFile));
+        assert(is_string($composerDir));
+        $this->composerDir = $composerDir;
+
+        $autoloadFilesPath = $composerDir . '/autoload_files.php';
+        if (! file_exists($autoloadFilesPath)) {
+            return;
+        }
+
+        /** @var array<string, string> $autoloadFiles */
+        $autoloadFiles = require $autoloadFilesPath;
+        foreach ($autoloadFiles as $autoloadFile) {
+            $path = $this->normalizePath($autoloadFile);
+            if (! is_string($path)) {
+                continue;
+            }
+
+            $this->composerLoadedFiles[$path] = true;
+        }
+    }
+
+    private function isComposerLoadedFile(string $filePath): bool
+    {
+        return str_starts_with($filePath, $this->composerDir . DIRECTORY_SEPARATOR)
+            || isset($this->composerLoadedFiles[$filePath]);
+    }
+
+    private function normalizePath(string $path): string|false
+    {
+        if (str_starts_with($path, 'phar://')) {
+            return $path;
+        }
+
+        return realpath($path);
+    }
+}
