+++ src/Context/CliModule.php
@@ -13,6 +13,7 @@ use BEAR\Sunday\Extension\Transfer\HttpCacheInterface;
 use BEAR\Sunday\Extension\Transfer\TransferInterface;
 use Override;
 use Ray\Di\AbstractModule;
+use Ray\Di\Name;
 
 use function crc32;
 use function sys_get_temp_dir;
@@ -26,7 +27,19 @@ final class CliModule extends AbstractModule
     #[Override]
     protected function configure(): void
     {
-        $this->rename(RouterInterface::class, 'original');
+        // Ray.Di 2.21+ defers rename() until after configure()+merge, so rename-then-bind
+        // in the same configure() renames the newly bound CliRouter. Move the parent
+        // binding on lastModule first (2.20 rename semantics), then bind CliRouter.
+        /** @psalm-suppress DeprecatedProperty lastModule is the chained parent module */
+        if ($this->lastModule instanceof AbstractModule) {
+            $this->lastModule->getContainer()->move(
+                RouterInterface::class,
+                Name::ANY,
+                RouterInterface::class,
+                'original',
+            );
+        }
+
         $this->bind(RouterInterface::class)->to(CliRouter::class);
         $this->bind(TransferInterface::class)->to(CliResponder::class);
         $this->bind(HttpCacheInterface::class)->to(CliHttpCache::class);
